pyproject.toml
[project]
name = "py-hse-monitoring"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "uvicorn>=0.34.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.20",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-decouple>=3.8",
    "bcrypt==4.3.0",
]
